% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log3models.R, R/log3agrt.R
\name{FitlogAGR}
\alias{FitlogAGR}
\title{Fits the three parameter logistic model for AGR}
\usage{
FitlogAGR(r, M, K, t)

FitlogAGR(r, M, K, t)
}
\arguments{
\item{r}{a number or numeric vector}

\item{M}{biomass, a number or numeric vector}

\item{K, }{upper horizontal asymptote, a number or numeric vector}

\item{t, }{time}

\item{r}{a number or numeric vector}

\item{M}{biomass, a number or numeric vector}

\item{K, }{upper horizontal asymptote, a number or numeric vector}

\item{t, }{time}
}
\value{
a numeric vector, Absolute Growth Rate (AGR) at time t

a numeric vector, Absolute Growth Rate (AGR) at time t
}

