% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log3agrt.R
\name{log.agrt}
\alias{log.agrt}
\title{Three parameter logistic}
\usage{
\method{log}{agrt}(r, M, K)
}
\arguments{
\item{r}{a number or numeric vector}

\item{M}{biomass, a number or numeric vector}

\item{K, }{upper horizontal asymptote, a number or numeric vector}

\item{t, }{time}
}
\value{
a numeric vector, AGR at time t for the 3 parameter logistic model
}

