% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log3dm.R
\name{log.dm}
\alias{log.dm}
\title{log.dm}
\usage{
\method{log}{dm}(r, M, K)
}
\arguments{
\item{r}{a number or numeric vector}

\item{M}{biomass, a number or numeric vector}

\item{K, }{upper horizontal asymptote, a number or numeric vector}
}
\value{
dM/dt, the change in mass over time or absolute growth rate (AGR) for the 3 parameter logistic model
}

