% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log3BM.R
\name{log.mass}
\alias{log.mass}
\title{log.mass}
\usage{
\method{log}{mass}(r, M, K)
}
\arguments{
\item{r}{a number or numeric vector}

\item{M}{biomass, a number or numeric vector}

\item{K, }{upper horizontal asymptote, a number or numeric vector}
}
\value{
a numeric vector, biomass at time t for 3 parametric logistic model
}

