% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log3rgrm.R
\name{log.rgrm}
\alias{log.rgrm}
\title{log.rgrm}
\usage{
\method{log}{rgrm}(r, M, K)
}
\arguments{
\item{r}{a number or numeric vector}

\item{M}{biomass, a number or numeric vector}

\item{K, }{upper horizontal asymptote, a number or numeric vector}
}
\value{
a numeric vector, RGR as a function of mass, 3 parameter logistic model
}

