% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log3rgrt.R
\name{log.rgrt}
\alias{log.rgrt}
\title{log.rgrt}
\usage{
\method{log}{rgrt}(r, M, K)
}
\arguments{
\item{r}{a number or numeric vector}

\item{M}{biomass, a number or numeric vector}

\item{K, }{upper horizontal asymptote, a number or numeric vector}

\item{t, }{time}
}
\value{
a numeric vector, relative growth rate as a function of time, 3 parameter logistic model
}

