% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log3dm.R
\name{Log3AGR}
\alias{Log3AGR}
\title{Three parameter logistic model for absolute growth rate}
\usage{
Log3AGR(r, M, K)
}
\arguments{
\item{r}{a constant, a number or numeric vector}

\item{M}{biomass, a number or numeric vector}

\item{K, }{upper horizontal asymptote, a number or numeric vector}
}
\value{
dM/dt, the change in mass over time or absolute growth rate (AGR) for the 3 parameter logistic model
}
\description{
\code{Log3AGR} returns the absolute growth rate (dM/dt) from the 3 parameter logistic model
}
\examples{
log.dmdt <- Log3AGR(0.25,100,1000)
log.dmdt
22.5


}
\references{
C.E.T. Paine, T.R. Marthews, D.R. Vogt, D. Purves, M. Rees, A. Hector, and
L.A. Turnbull. 2012. How to fit nonlinear growth models and calculate growth
rates: an update for ecologists. Methods in Ecology and Evolution 3:245-256
}

