% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log3models.R
\name{Log3AGRt}
\alias{Log3AGRt}
\title{Three parameter logistic AGR time basis}
\usage{
Log3AGRt(r, M, K, t)
}
\arguments{
\item{r}{a constant, a number or numeric vector}

\item{M}{initial biomass, a number or numeric vector}

\item{K}{upper horizontal asymptote, a number or numeric vector}

\item{t}{time}
}
\value{
a numeric vector, Absolute Growth Rate (AGR) at time t
}
\description{
\code{Log3AGRt} returns returns the absolute growth rate on a time basis for cross species comparisons
  for the 3 parameter logistic model
}
\examples{
#Time basis AGR after 1 time step
agr.t <- Log3AGRt(0.25,100,1000,1)
agr.t
27.31

}
\references{
C.E.T. Paine, T.R. Marthews, D.R. Vogt, D. Purves, M. Rees, A. Hector, and
L.A. Turnbull. 2012. How to fit nonlinear growth models and calculate growth
rates: an update for ecologists. Methods in Ecology and Evolution 3:245-256
}

