% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log3rgrm.R
\name{Log3RGRm}
\alias{Log3RGRm}
\title{Relative growth rate as a function of mass for the 3 parameter logistic model}
\usage{
Log3RGRm(r, M, K)
}
\arguments{
\item{r}{a constant, a number or numeric vector}

\item{M}{initial biomass, a number or numeric vector}

\item{K}{upper horizontal asymptote, a number or numeric vector}
}
\value{
a numeric vector, RGR as a function of mass, 3 parameter logistic model
}
\description{
\code{Log3RGRm} returns the relative growth rate as a function of mass for the 3 parameter logistic model
}
\examples{
#RGR as a function of mass 
rgr.mass <- Log3RGRm(0.25,100,1000)
rgr.mass
0.225

}
\references{
C.E.T. Paine, T.R. Marthews, D.R. Vogt, D. Purves, M. Rees, A. Hector, and
L.A. Turnbull. 2012. How to fit nonlinear growth models and calculate growth
rates: an update for ecologists. Methods in Ecology and Evolution 3:245-256
}

