% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log3rgrt.R
\name{Log3RGRt}
\alias{Log3RGRt}
\title{Relative growth rate as a function of time for the 3 parameter logistic model}
\usage{
Log3RGRt(r, M, K, t)
}
\arguments{
\item{r}{a constant, a number or numeric vector}

\item{M}{initial biomass, a number or numeric vector}

\item{K}{upper horizontal asymptote, a number or numeric vector}

\item{t}{time, a number or numeric vector}
}
\value{
a numeric vector, relative growth rate as a function of time, 3 parameter logistic model
}
\description{
\code{Log3RGRt} returns the relative growth rate as a function of time for the 3 parameter logistic
}
\examples{
RGR after one time step
rgr.t <- Log3RGRt(0.25,100,1000,1)
rgr.t
0.2187

}
\references{
C.E.T. Paine, T.R. Marthews, D.R. Vogt, D. Purves, M. Rees, A. Hector, and
L.A. Turnbull. 2012. How to fit nonlinear growth models and calculate growth
rates: an update for ecologists. Methods in Ecology and Evolution 3:245-256
}

