% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoAGRt.R
\name{mono.agrt}
\alias{mono.agrt}
\title{mono.agrt}
\usage{
mono.agrt(r, M, K, t)
}
\arguments{
\item{r}{growth rate, a number or numeric vector}

\item{M}{M0 initial biomass, a number or numeric vector}

\item{K}{upper horizontal asymptote, a number or numeric vector}

\item{t}{time, a number or numeric vector}
}
\value{
AGR as a function of time for the monomolecular model
}

