% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoBM.R
\name{mono.mass}
\alias{mono.mass}
\title{mono.mass}
\usage{
mono.mass(r, M, K, t)
}
\arguments{
\item{r}{a number or numeric vector}

\item{M}{M0 initial biomass, a number or numeric vector}

\item{K, }{upper horizontal asymptote, a number or numeric vector}

\item{t, }{time, a number or numeric vector}
}
\value{
Biomass at time t for the monomolecular model of growth
}

