% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoRGRm.R
\name{mono.rgrm}
\alias{mono.rgrm}
\title{mono.rgrm}
\usage{
mono.rgrm(r, M, K)
}
\arguments{
\item{r}{a number or numeric vector}

\item{M}{M0 initial biomass, a number or numeric vector}

\item{K, }{upper horizontal asymptote, a number or numeric vector}
}
\value{
RGR as a function of mass
}

