% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoRGRt.R
\name{mono.rgrt}
\alias{mono.rgrt}
\title{mono.rgrt}
\usage{
mono.rgrt(r, M, K, t)
}
\arguments{
\item{r}{a number or numeric vector}

\item{M}{M0 initial biomass, a number or numeric vector}

\item{K, }{upper horizontal asymptote, a number or numeric vector}

\item{t, }{time, a number or numeric vector}
}
\value{
RGR as a function of time
}

