% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoAGR.R
\name{MonoAGR}
\alias{MonoAGR}
\title{Monomolecular function for absolute growth rate}
\usage{
MonoAGR(r, M, K)
}
\arguments{
\item{r}{a constant, a number or numeric vector}

\item{M}{biomass, a number or numeric vector}

\item{K}{upper horizontal asymptote, a number or numeric vector}
}
\value{
dM/dt, the change in mass over time or absolute growth rate (AGR) for the 3 parameter logistic model
}
\description{
\code{MonoAGR} returns the absolute growth rate (dM/dt)
}
\references{
C.E.T. Paine, T.R. Marthews, D.R. Vogt, D. Purves, M. Rees, A. Hector, and
L.A. Turnbull. 2012. How to fit nonlinear growth models and calculate growth
rates: an update for ecologists. Methods in Ecology and Evolution 3:245-256
}

