% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoAGRt.R
\name{MonoAGRt}
\alias{MonoAGRt}
\title{Monomolecular AGR time basis}
\usage{
MonoAGRt(r, M, K, t)
}
\arguments{
\item{r}{allometric constant, a number or numeric vector}

\item{M}{initial biomass (kg/g), a number or numeric vector}

\item{K}{upper asymptote, a number or numeric vector}

\item{t}{time, a number or numeric vector}
}
\value{
The absolute growth rate as a function of time
}
\description{
\code{MonoAGRt}  returns the absolute growth rate on a time basis for cross species comparisons
  following the monomolecular function
}
\references{
C.E.T. Paine, T.R. Marthews, D.R. Vogt, D. Purves, M. Rees, A. Hector, and
L.A. Turnbull. 2012. How to fit nonlinear growth models and calculate growth
rates: an update for ecologists. Methods in Ecology and Evolution 3:245-256
}

