% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monodm.R
\name{monom}
\alias{monom}
\title{monom    
Model of change is mass over time dM/dt}
\usage{
monom(r, M, K)
}
\arguments{
\item{r}{a number or numeric vector}

\item{M}{biomass, a number or numeric vector}

\item{K, }{upper horizontal asymptote, a number or numeric vector}
}
\value{
dM/dt, the change in mass over time or absolute growth rate (AGR) for the monomolecular model
}

