% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoBM.R
\name{MonoMass}
\alias{MonoMass}
\title{Mass at time t following the monomolecular function}
\usage{
MonoMass(r, M, K, t)
}
\arguments{
\item{r}{allometric constant, a number or numeric vector}

\item{M}{intial biomass, a number or numeric vector}

\item{K}{upper horizontal asymptote, a number or numeric vector}

\item{t}{time a number or numeric vector}
}
\value{
number or numeric vector, biomass at time t from the power law function
}
\description{
\code{MonoMass} returns biomass at time t
}
\references{
C.E.T. Paine, T.R. Marthews, D.R. Vogt, D. Purves, M. Rees, A. Hector, and
L.A. Turnbull. 2012. How to fit nonlinear growth models and calculate growth
rates: an update for ecologists. Methods in Ecology and Evolution 3:245-256
}

