% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoRGRm.R
\name{MonoRGRm}
\alias{MonoRGRm}
\title{Relative growth rate as a function of mass}
\usage{
MonoRGRm(r, M, K)
}
\arguments{
\item{r}{a constant, a number or numeric vector}

\item{M}{initial biomass, a number or numeric vector}

\item{K}{upper horizontal asymptote, a number or numeric vector}
}
\value{
a numeric vector, RGR as a function of mass
}
\description{
\code{MonoRGRm} returns the relative growth rate as a function of mass for the monomolecular function
}
\references{
C.E.T. Paine, T.R. Marthews, D.R. Vogt, D. Purves, M. Rees, A. Hector, and
L.A. Turnbull. 2012. How to fit nonlinear growth models and calculate growth
rates: an update for ecologists. Methods in Ecology and Evolution 3:245-256
}

