% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoRGRt.R
\name{MonoRGRt}
\alias{MonoRGRt}
\title{Relative growth rate as a function of time}
\usage{
MonoRGRt(r, M, K, t)
}
\arguments{
\item{r}{a constant, a number or numeric vector}

\item{M}{initial biomass, a number or numeric vector}

\item{K}{upper horizontal asymptote, a number or numeric vector}

\item{t}{time, a number or numeric vector}
}
\value{
a number or numeric vector, the relative growth rate at time t
}
\description{
\code{MonoRGRt} returns the relative growth rate at time t for the monomolecular function
}
\references{
C.E.T. Paine, T.R. Marthews, D.R. Vogt, D. Purves, M. Rees, A. Hector, and
L.A. Turnbull. 2012. How to fit nonlinear growth models and calculate growth
rates: an update for ecologists. Methods in Ecology and Evolution 3:245-256
}

