% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerAGRt.R
\name{pl.agrt}
\alias{pl.agrt}
\title{p.agrt}
\usage{
pl.agrt(r, M, B, t)
}
\arguments{
\item{r}{a number or numeric vector}

\item{M}{intial biomass, a number or numeric vector}

\item{B}{a number or numeric vector}

\item{t}{time}
}
\value{
The absolute growth rate as a function of time
}

