% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerBM.R
\name{pl.mass}
\alias{pl.mass}
\title{pl.mass}
\usage{
pl.mass(r, M, B, t)
}
\arguments{
\item{r}{a number or numeric vector}

\item{M}{intial biomass a number or numeric vector}

\item{B}{a number or numeric vector}

\item{t}{time}
}
\value{
Biomass at time t for the power law/allometric model
}

