% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerRGRm.R
\name{pl.rgrm}
\alias{pl.rgrm}
\title{pl.rgrm

Relative growth rate as a function of mass following the power law}
\usage{
pl.rgrm(r, M, B)
}
\arguments{
\item{r}{a number or numeric vector}

\item{M}{intial biomass a number or numeric vector}

\item{B}{a number or numeric vector}
}
\value{
The relative growth rate as a function of mass
}

