% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerRGR.R
\name{pl.rgt}
\alias{pl.rgt}
\title{pl.rgrt}
\usage{
pl.rgt(r, M, B, t)
}
\arguments{
\item{r}{a number or numeric vector}

\item{M}{biomass a number or numeric vector}

\item{B}{a number or numeric vector}

\item{t}{time a number or numeric vector}
}
\value{
The relative growth rate at time t following the power law
}

