% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw.R
\name{powerlaw}
\alias{powerlaw}
\title{powerlaw}
\usage{
powerlaw(r, M, B)
}
\arguments{
\item{r}{a number or numeric vector}

\item{M}{biomass a number or numeric vector}

\item{B}{a number or numeric vector}
}
\value{
The absolute growth rate following the powerlaw form rM^B
}

