% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw.R
\name{PowerlawAGR}
\alias{PowerlawAGR}
\title{Powerlaw function for absolute growth rate (change in mass over time)}
\usage{
PowerlawAGR(r, M, B)
}
\arguments{
\item{r}{a number or numeric vector}

\item{M}{biomass a number or numeric vector}

\item{B}{a number or numeric vector}
}
\value{
numeric vector of the absolute growth rate following the powerlaw form rM^B
}
\description{
\code{PowerlawAGR}  returns the absolute growth rate following a powerlaw/allometric functional form
}

