% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerBM.R
\name{PowerlawMass}
\alias{PowerlawMass}
\title{Mass at time t following the powerlaw function (allometric growth)}
\usage{
PowerlawMass(r, M, B, t)
}
\arguments{
\item{r}{a number or numeric vector}

\item{M}{intial biomass a number or numeric vector}

\item{B}{a number or numeric vector}

\item{t}{time a number or numeric vector}
}
\value{
numeric vector, Biomass at time t from the powerlaw function
}
\description{
\code{PowerlawMass} returns biomass at time t from the powerlaw function
}

