% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerRGRm.R
\name{PowerlawRGRm}
\alias{PowerlawRGRm}
\title{Relative growth rate as a function of mass}
\usage{
PowerlawRGRm(r, M, B)
}
\arguments{
\item{r}{a number or numeric vector}

\item{M}{intial biomass a number or numeric vector}

\item{B}{a number or numeric vector}
}
\value{
The relative growth rate as a function of mass
}
\description{
\code{PowerlawRGRm} returns the relative growth rate as a function of mass from the powerlaw function
}

