% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw.R
\name{PowerlawAGR}
\alias{PowerlawAGR}
\title{Powerlaw function for absolute growth rate (change in mass over time)}
\usage{
PowerlawAGR(r, M, B)
}
\arguments{
\item{r}{allometric constant, a number or numeric vector}

\item{M}{biomass (kg/g), a number or numeric vector}

\item{B}{scaling exponent, a number or numeric vector}
}
\value{
a number or numeric vector of the absolute growth rate following the power law
}
\description{
\code{PowerlawAGR}  returns the absolute growth rate following a power law/allometric functional form
}
\examples{
agr <- PowerlawAGR(0.25,100,0.4)
#Change in mass over time
agr <- 3.62 

}
\references{
C.E.T. Paine, T.R. Marthews, D.R. Vogt, D. Purves, M. Rees, A. Hector, and
L.A. Turnbull. 2012. How to fit nonlinear growth models and calculate growth
rates: an update for ecologists. Methods in Ecology and Evolution 3:245-256
}

