% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerRGRm.R
\name{PowerlawRGRm}
\alias{PowerlawRGRm}
\title{Relative growth rate as a function of mass following the power law/allometric form}
\usage{
PowerlawRGRm(r, M, B)
}
\arguments{
\item{r}{allometric constant, a number or numeric vector}

\item{M}{intial biomass (kg/g), a number or numeric vector}

\item{B}{scaling exponent, a number or numeric vector}
}
\value{
a number or numeric vector, the relative growth rate as a function of mass
}
\description{
\code{PowerlawRGRm} returns the relative growth rate as a function of mass from the powerlaw function
}
\examples{
#RGR after one time step 
rgr.m <- PowerlawRGRm(0.25,100,0.4)
rgr.m 
0.57739

}
\references{
C.E.T. Paine, T.R. Marthews, D.R. Vogt, D. Purves, M. Rees, A. Hector, and
L.A. Turnbull. 2012. How to fit nonlinear growth models and calculate growth
rates: an update for ecologists. Methods in Ecology and Evolution 3:245-256
}

