% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerRGR.R
\name{PowerlawRGRt}
\alias{PowerlawRGRt}
\title{Relative growth rate as a function of time following the powerlaw/allometric form}
\usage{
PowerlawRGRt(r, M, B, t)
}
\arguments{
\item{r}{a number or numeric vector}

\item{M}{biomass a number or numeric vector}

\item{B}{a number or numeric vector}

\item{t}{time a number or numeric vector}
}
\value{
The relative growth rate at time t following the powerlaw
}
\description{
\code{PowerlawRGRt} returns the relative growth rate at any time t for the powerlaw/allometric functional form
}

